refactor configurations
