+++ src/assets/image.js
@@ -1,12 +1,12 @@
 import _ from 'lodash'
-import {CDN_URL} from '@/lib/config'
+import {CDN_URL} from '@/config/const'
 
 export const withRatio = (dimensions) => (options) => {
   const style = _.defaults({}, options, {position: 'relative'}, dimensions)
   const {width, height, resizeMode} = options
   if (width && height) style.resizeMode = resizeMode || 'cover'
-  else if (height) style.width = dimensions.width * height / dimensions.height
-  else if (width) style.height = dimensions.height * width / dimensions.width
+  else if (height) style.width = (dimensions.width * height) / dimensions.height
+  else if (width) style.height = (dimensions.height * width) / dimensions.width
   return style
 }
