+++ src/config/tabs.js
@@ -3,6 +3,13 @@ import * as listings from '@/screens/modules/listings/screens'
 import * as account from '@/screens/modules/account/screens'
 import * as auth from '@/screens/modules/auth/screens'
 
+/**
+ * react-native-navigation bottom tabs layout generator.
+ * Returns an array of `component` layouts for each tab.
+ * https://wix.github.io/react-native-navigation/v2/#/docs/layout-types?id=component
+ * @param {Object} state - redux state
+ * @param {Object|null} args.user - logged in user
+ */
 export default (_, {user}) => [
   {
     name: listings.Feed.screenName,
