+++ src/graphql/containers/SendMessageMutation.js
@@ -1,6 +1,6 @@
 import {Mutation} from 'react-apollo'
 
-import {MESSENGER_RECEIVER_ID} from '@/lib/config'
+import {MESSENGER_RECEIVER_ID} from '@/config/const'
 import {
   GET_MESSAGES,
   GET_CHANNEL_FEED
