+++ src/graphql/links/wsLink/index.js
@@ -1,5 +1,5 @@
 import WSLink from './Link'
 
-import {WEB_SOCKET_URL} from '@/lib/config'
+import {WEB_SOCKET_URL} from '@/config/const'
 
 export default () => new WSLink(WEB_SOCKET_URL)
