+++ src/redux/modules/reducer.js
@@ -3,7 +3,7 @@ import {combineReducers} from 'redux'
 import {persistReducer} from 'redux-persist'
 import {reducer as network} from 'react-native-offline'
 
-import {PERSIST_TIMEOUT} from '@/lib/config'
+import {PERSIST_TIMEOUT} from '@/config/const'
 import auth from './auth'
 import relatedListings from './relatedListings'
 import gallery from './gallery'
