+++ src/screens/modules/listings/reducer.js
@@ -2,7 +2,7 @@ import {combineReducers} from 'redux'
 import {persistReducer} from 'redux-persist'
 import {AsyncStorage} from 'react-native'
 
-import {PERSIST_TIMEOUT} from '@/lib/config'
+import {PERSIST_TIMEOUT} from '@/config/const'
 import Map from './Map/module'
 import Search from './Search/module'
